+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/TeacherRestController.java
@@ -2,8 +2,11 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.TeacherDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.TeacherController;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/teachers")
@@ -24,4 +27,17 @@ public class TeacherRestController {
     public Teacher updateTeacher(@PathVariable Long teacherId, @RequestBody TeacherDTO teacherDTO){
         return teacherController.updateTeacher(teacherId, teacherDTO);
     }
+
+    @GetMapping("/{teacherId}")
+    public Teacher getTeacherById(@PathVariable Long teacherId){
+        return teacherController.getTeacherById(teacherId);
+    }
+
+    @GetMapping ("courses/{teacherId}")
+    public List<Course> getTeacherCourses(@PathVariable Long teacherId){
+        return teacherController.getTeacherCourses(teacherId);
+    }
+
+
+
 }