+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [NPP-26] 》TE18F - Create a teacher  
 - [NPP-27] 》TE19F - Update a teacher by ID
+- [NPP-29] 》TE21F - Get all courses from a teacher by teacher ID
 
 
 - [NPP-44] 》ST38F - Create a student associated with a course